move icon png up a directory
I previously had goggles.png in a subdirectory of the package dir.
This was because the UtilServer was failing to expose it at the
package level for reasons I didn't understand.

Kai pointed out why this was - the wrong classloader was being used
(actually the UtilServer.addLocalUrl(String) javadoc gives a heavy
hint that this could be a problem, but I missed it).
Work round that by avoiding that utility method.

I will also deprecate that, somewhat dangerous, method in JSAMP.